Add some refactoring changes (2)
