+++ ui/components/app/snaps/snap-permission-cell-display/index.js
@@ -1,1 +1,1 @@
-export { default } from './snap-permission-cell-display';
+export { default } from './snap-permission-cell';
